use float instead of double

+++ include/frugally_deep/typedefs.hpp
@@ -14,7 +14,7 @@
 namespace fdeep { namespace internal
 {
 
-typedef double float_t;
+typedef float float_t;
 typedef std::vector<float_t> float_vec;
 typedef float_vec::const_iterator float_vec_const_it;
 typedef float_vec::iterator float_vec_it;
